+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
 ### Added
+- **Experimental** A ``pw.xpacks.llm.document_store.DocumentStore`` to process and index documents.
 - `pw.xpacks.stdlib.indexing.HybridIndex` used for querying multiple indices and combining their results.
 
 ### Changed